xilpm: versal: server: Fix MISRA Rule 10.3
This patch fixes MISRA C-2012 Rule 10.3, i.e. The value of an
expression shall not be assigned to an object with a narrower
essential type or a different essential type category.


Acked-for-series: Rajan Vaja <rajan.vaja@xilinx.com>

+++ lib/sw_services/xilpm/src/versal/server/xpm_bisr.c
@@ -1250,8 +1250,8 @@ static u32 XPmBisr_RepairLaguna(u32 EfuseTagAddr, u32 TagSize)
 		Row1 = (u8)(LagunaRepairY1 / 96U);
 
 		/* Get tile index within FSR row */
-		Tile0 = LagunaRepairY0 % 96U;
-		Tile1 = LagunaRepairY1 % 96U;
+		Tile0 = (u8)(LagunaRepairY0 % 96U);
+		Tile1 = (u8)(LagunaRepairY1 % 96U);
 
 		/* Walk through each FSR row that requires repair */
 		for (RowIndex = Row0; RowIndex <= Row1; RowIndex++) {
